8.15

+++ app/src/main/java/com/example/pmp/data/model/CreateProjectData.kt
@@ -4,5 +4,5 @@ data class CreateProjectData(
     val name: String,
     val description: String,
     val isPublic: Boolean,
-    val userId: Int,
+    val userId: Long
 )
